+++ src/language/helpers.ts
@@ -1,7 +1,13 @@
 import { ParameterType, RegExps } from '@cucumber/cucumber-expressions'
 import { DocumentUri, LocationLink, Range } from 'vscode-languageserver-types'
 
-import { Link, NodePredicate, TreeSitterQueryMatch, TreeSitterSyntaxNode } from './types'
+import {
+  LanguageName,
+  Link,
+  NodePredicate,
+  TreeSitterQueryMatch,
+  TreeSitterSyntaxNode,
+} from './types'
 
 export function syntaxNode(match: TreeSitterQueryMatch, name: string): TreeSitterSyntaxNode | null {
   const nodes = syntaxNodes(match, name)
@@ -69,4 +75,67 @@ export function filter(
 
 function flatten(node: TreeSitterSyntaxNode): TreeSitterSyntaxNode[] {
   return node.children.reduce((r, o) => [...r, ...flatten(o)], [node])
+}
+
+/**
+ *
+ * This constant represents a record of language names that contain lists
+ * of regular expressions that should be stripped from the content of a file.
+ */
+export const BLACKLISTED_EXPRESSIONS: {
+  [key in LanguageName]: RegExp[]
+} = {
+  tsx: [
+    /*
+     * This regular expression matches sequences of decorators applied to a class,
+     * potentially including type parameters and arguments.
+     * The regex supports matching these patterns preceding
+     * an optionally exported class definition.
+     */
+    /(@(\w+)(?:<[^>]+>)?\s*(?:\([^)]*\))?\s*)*(?=\s*(export)*\s+class)/g,
+  ],
+  java: [],
+  c_sharp: [],
+  php: [],
+  python: [],
+  ruby: [],
+  rust: [],
+  javascript: [],
+}
+
+/**
+ *
+ * Strips blacklisted expressions from the given content.
+ *
+ * @param content The content to strip blacklisted expressions from.
+ * @param languageName The name of the language to use for stripping.
+ *
+ * @returns The content with blacklisted expressions stripped.
+ *
+ * @example
+ *
+ * ```typescript
+ * const content =
+ *  "@decorator\n" ++
+ *  "export class Foo {\n" ++
+ *  "@decorator\n" ++
+ *  "public bar() { }\n" ++
+ *  "}"
+ *
+ * const strippedContent = stripBlacklistedExpressions(content, 'tsx')
+ * console.log(strippedContent)
+ *
+ * // Output:
+ * "export class Foo {\n" ++
+ * "@decorator\n" ++
+ * "public bar() { }\n" ++
+ * "}"
+ *
+ * ```
+ */
+export function stripBlacklistedExpressions(content: string, languageName: LanguageName): string {
+  return BLACKLISTED_EXPRESSIONS[languageName].reduce(
+    (acc, regExp) => acc.replace(regExp, ''),
+    content
+  )
 }