+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - (Rust) Support for r# raw strings with step definition patterns ([#176](https://github.com/cucumber/language-service/pull/176))
 - (Rust) Line continuation characters in rust step definition patterns ([#179](https://github.com/cucumber/language-service/pull/179))
 - (Python) Unexpected spaces and commas in generated step definitions [#160](https://github.com/cucumber/language-service/issues/160)
+- (TypeScript) Tree sitter parser failing on class decorators ([#186](https://github.com/cucumber/language-service/pull/186))
 
 ### Added
 - (Python) Support for u-strings with step definition patterns ([#173](https://github.com/cucumber/language-service/pull/173))